refactor: compose detail header actions dropdown
