+++ web/src/features/traces/components/TraceDetailView/components/TraceDetailViewHeader.tsx
@@ -21,7 +21,7 @@ import { type WithStringifiedMetadata } from "@/src/utils/clientSideDomainTypes"
 import { type ObservationReturnTypeWithMetadata } from "@/src/server/api/routers/traces";
 import { ItemBadge } from "@/src/components/ItemBadge";
 import { LocalIsoDate } from "@/src/components/LocalIsoDate";
-import { DetailHeaderActionsMenu } from "@/src/features/traces/components/DetailHeaderActionsMenu";
+import { DetailHeaderActionsMenuController } from "@/src/features/traces/components/DetailHeaderActionsMenuController";
 import { NewDatasetItemFromExistingObject } from "@/src/features/datasets/components/NewDatasetItemFromExistingObject";
 import { AnnotateDrawer } from "@/src/features/scores/components/AnnotateDrawer";
 import { AnnotationQueueItemDropdownMenuController } from "@/src/features/annotation-queues/components/AnnotationQueueItemDropdownMenuController";
@@ -43,7 +43,12 @@ import { useViewPreferences } from "@/src/features/traces/contexts/ViewPreferenc
 import { CollapsibleBadgeRow } from "@/src/features/traces/components/CollapsibleBadgeRow";
 import { useIsMobile } from "@/src/hooks/use-mobile";
 import { Button } from "@/src/components/ui/button";
-import { ChevronDown, ListPlus, MoreHorizontal } from "lucide-react";
+import {
+  ChevronDown,
+  EllipsisVertical,
+  ListPlus,
+  MoreHorizontal,
+} from "lucide-react";
 import {
   Popover,
   PopoverContent,
@@ -107,14 +112,28 @@ export const TraceDetailViewHeader = memo(function TraceDetailViewHeader({
           >
             {trace.name || trace.id}
           </span>
-          <DetailHeaderActionsMenu
+          <DetailHeaderActionsMenuController
             idItems={[{ id: trace.id, name: "Trace ID" }]}
             projectId={projectId}
             webCallout={{
               traceId: trace.id,
               sessionId: trace.sessionId ?? null,
             }}
-          />
+          >
+            {({ Trigger }) => (
+              <Trigger asChild>
+                <Button
+                  aria-label="Options"
+                  className="mt-0.5 shrink-0"
+                  size="icon-xs"
+                  title="Options"
+                  variant="ghost"
+                >
+                  <EllipsisVertical className="h-4 w-4" />
+                </Button>
+              </Trigger>
+            )}
+          </DetailHeaderActionsMenuController>
           {/* Mobile: collapse the action-button cluster into a `⋯` overflow of
               full-width labeled rows, next to the `⋮` utility menu. */}
           {isMobile && (
